fix(v1.41.0): conversations.user_id is BIGINT to match users.id (BIGSERIAL)
DBA reviewer flagged: every other FK to users.id in the schema uses BIGINT
(infra/init/07, 08, 10, 12, 13). INTEGER would overflow once user IDs cross
2^31. Bring this in line before the table is populated.

+++ infra/init/38_copilot_conversations.sql
@@ -6,7 +6,8 @@
 
 CREATE TABLE IF NOT EXISTS conversations (
     id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    user_id         INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    -- users.id is BIGSERIAL (05_users.sql); FK must match the parent type.
+    user_id         BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     workspace_id    UUID REFERENCES workspaces(id) ON DELETE CASCADE,
     title           TEXT,
     created_at      TIMESTAMPTZ NOT NULL DEFAULT now(),
